tests: corrige testes de usuario

+++ src/application/useCase/post/deleteById/delete.ts
@@ -1,20 +1,21 @@
 import { PostAggregateRepositoryInterface } from '../../../../domain/post/repository/postRepository.interface';
-import { AppError } from '../../../../infrastructure/api/errors';
-import { errorStates } from '../../../../infrastructure/api/errors/types';
+import { AppError } from '../../../errors/AppError';
+import { DeletePostUseCaseInterface } from './DeletePostUseCaseInterface';
 
-export class DeletePostUseCase {
+export class DeletePostUseCase implements DeletePostUseCaseInterface {
   constructor(private postRepository: PostAggregateRepositoryInterface) {}
 
   execute = async (idPost: string, userId: string): Promise<void> => {
     const post = await this.postRepository.findById(idPost);
-    if (!post?.userId || !userId || post?.userId !== userId) {
-      // você não pode deletar posts que não sejam seus
-      throw new AppError(errorStates.FORBIDDEN);
+    if (!post) {
+      throw new AppError('POST_NOT_EXISTS');
     }
 
-    const postDeleted = this.postRepository.deleteById(idPost);
-    if (postDeleted === null) {
-      throw new AppError(errorStates.RESOURCE_NOT_EXISTS);
+    if (!post?.userId || !userId || post?.userId !== userId) {
+      // você não pode deletar posts que não sejam seus
+      throw new AppError('NO_CAN_DELETE_POST_ANOTHER_USER');
     }
+
+    this.postRepository.deleteById(idPost);
   };
 }
